backend/scripts/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.18.1
webdriver-manager==4.0.1 
python-dotenv